refactor(domain.Payment) : 매개 변수 amount의 타입이 int였던 오류 수정

+++ src/main/java/lotto/domain/Payment.java
@@ -6,38 +6,38 @@ import lotto.constant.LottoConstraint;
 public class Payment {
     private final long amount;
 
-    public Payment(int amount) {
+    public Payment(long amount) {
         validate(amount);
         this.amount = amount;
     }
 
-    private void validate(int amount) {
+    private void validate(long amount) {
         validateDivisibleAmount(amount);
         validateAmountRange(amount);
     }
 
-    private void validateDivisibleAmount(int amount) {
+    private void validateDivisibleAmount(long amount) {
         if (isIndivisibleAmount(amount)) {
             throw new IllegalArgumentException(ErrorMessage.INDIVISIBLE_AMOUNT.getMessage());
         }
     }
 
-    private boolean isIndivisibleAmount(int amount) {
+    private boolean isIndivisibleAmount(long amount) {
         return amount % LottoConstraint.PRICE_PER_LOTTO.getValue() != 0;
     }
 
-    private void validateAmountRange(int amount) {
+    private void validateAmountRange(long amount) {
         if (isOutRangeAmount(amount)) {
             throw new IllegalArgumentException(ErrorMessage.OUT_RANGE_AMOUNT.getMessage());
         }
     }
 
-    private boolean isOutRangeAmount(int amount) {
+    private boolean isOutRangeAmount(long amount) {
         return amount < LottoConstraint.MIN_AMOUNT.getValue() || amount > LottoConstraint.MAX_AMOUNT.getValue();
     }
 
     public int calculatePurchaseLottoCount() {
-        return (int) amount / LottoConstraint.PRICE_PER_LOTTO.getValue();
+        return (int) (amount / LottoConstraint.PRICE_PER_LOTTO.getValue());
     }
 
     long getAmount() {
